COWER-46: Split backend solution in projects (#15)
* COWER-46: Split solution

* Fix dockerfile

* Add domain project

* Use AppRoles class instead of magic number
