+++ app/backend/Cower.Data/ApplicationContext.cs
@@ -1,0 +1,20 @@
+using Cower.Data.Entities;
+using Microsoft.EntityFrameworkCore;
+
+namespace Cower.Data;
+
+public sealed class ApplicationContext : DbContext
+{
+    public DbSet<UserEntity> Users { get; set; }
+    public DbSet<RoleEntity> Roles { get; set; }
+    
+    public ApplicationContext()
+    {
+        Database.EnsureCreated();
+    }
+    
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+    {
+        optionsBuilder.UseNpgsql(Environment.GetEnvironmentVariable("DB_CONNECTION_STRING"));
+    }
+}
