+++ app/backend/Cower.Data/Entities/RoleEntity.cs
@@ -1,0 +1,7 @@
+namespace Cower.Data.Entities;
+
+public class RoleEntity
+{
+    public long Id { get; set; }
+    public string Name { get; set; }
+}
