+++ app/backend/Cower.Data/Entities/UserEntity.cs
@@ -1,0 +1,17 @@
+using Microsoft.EntityFrameworkCore;
+
+namespace Cower.Data.Entities;
+
+[Index(nameof(Email), IsUnique = true)]
+public class UserEntity
+{
+    public long Id { get; set; }
+    public string Email { get; set; }
+    public byte[] PasswordHash { get; set; }
+    public string? Name { get; set; }
+    public string? Surname { get; set; }
+    public string? Phone { get; set; }
+    public long RoleId { get; set; }
+
+    public RoleEntity Role { get; set; }
+}
