+++ app/backend/Cower.Data/Repositories/IUserRepository.cs
@@ -1,0 +1,10 @@
+using Cower.Data.Entities;
+
+namespace Cower.Data.Repositories;
+
+public interface IUserRepository
+{
+    public UserEntity? GetUser(long id);
+    public UserEntity? GetUserByCredentials(string email, byte[] password);
+    public UserEntity AddUser(UserEntity user);
+}
