+++ app/backend/Cower.Domain/Models/Roles/AppRoles.cs
@@ -1,0 +1,7 @@
+namespace Cower.Domain.Models;
+
+public static class AppRoles
+{
+    public static Role Admin = new Role("Admin", 1);
+    public static Role User = new Role("User", 2);
+}
