+++ app/backend/Cower.Domain/Models/Roles/Role.cs
@@ -1,0 +1,5 @@
+namespace Cower.Domain.Models;
+
+public record Role(
+    string Name,
+    long Id);
