+++ app/backend/Cower.Domain/Models/User.cs
@@ -1,0 +1,10 @@
+namespace Cower.Domain.Models;
+
+public record User(
+    long Id,
+    string Email,
+    byte[] PasswordHash,
+    string? Name,
+    string? Surname,
+    string? Phone,
+    Role Role);
