+++ app/backend/Cower.Service/Models/RegisterUserRequestBL.cs
@@ -1,0 +1,8 @@
+namespace Cower.Service.Models;
+
+public record RegisterUserRequestBL(
+    string Email,
+    string Password,
+    string? Name,
+    string? Surname,
+    string? Phone);
