+++ app/backend/Cower.Service/Services/IJwtService.cs
@@ -1,0 +1,8 @@
+using Cower.Domain.Models;
+
+namespace Cower.Service.Services;
+
+public interface IJwtService
+{
+    public string GenerateJwt(User user);
+}
