+++ app/backend/Cower.Web/Models/LoginRequestDTO.cs
@@ -1,0 +1,11 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace Cower.Web.Models;
+
+public class LoginRequestDTO
+{
+    [Required]
+    public string Email { get; set; }
+    [Required]
+    public string Password { get; set; }
+}
