+++ app/backend/Cower.Web/Models/LoginResponseDTO.cs
@@ -1,0 +1,5 @@
+namespace Cower.Web.Models;
+
+public record LoginResponseDTO(
+    UserResponseDTO User,
+    string Jwt);
