+++ app/backend/Cower.Web/Models/RegisterResponseDTO.cs
@@ -1,0 +1,5 @@
+namespace Cower.Web.Models;
+
+public record RegisterResponseDTO(
+    UserResponseDTO User,
+    string Jwt);
