+++ app/backend/Cower.Web/Models/UserInfoResponseDTO.cs
@@ -1,0 +1,3 @@
+namespace Cower.Web.Models;
+
+public record UserInfoResponseDTO(UserResponseDTO User);
