+++ app/backend/Cower.Web/Models/UserResponseDTO.cs
@@ -1,0 +1,9 @@
+namespace Cower.Web.Models;
+
+public record UserResponseDTO(
+    long Id,
+    string Email,
+    string Role,
+    string? Name,
+    string? Surname,
+    string? Phone);
