--- app/backend/Cower/Data/ApplicationContext.cs
@@ -1,20 +1,0 @@
-using Cower.Data.Models;
-using Microsoft.EntityFrameworkCore;
-
-namespace Cower.Data;
-
-public sealed class ApplicationContext : DbContext
-{
-    public DbSet<User> Users { get; set; }
-    public DbSet<Role> Roles { get; set; }
-    
-    public ApplicationContext()
-    {
-        Database.EnsureCreated();
-    }
-    
-    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-    {
-        optionsBuilder.UseNpgsql(Environment.GetEnvironmentVariable("DB_CONNECTION_STRING"));
-    }
-}
