--- app/backend/Cower/Data/Models/Role.cs
@@ -1,7 +1,0 @@
-namespace Cower.Data.Models;
-
-public class Role
-{
-    public long Id { get; set; }
-    public string Name { get; set; }
-}
