--- app/backend/Cower/Presentation/Models/LoginRequestDTO.cs
@@ -1,11 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace Cower.Presentation.Models;
-
-public class LoginRequestDTO
-{
-    [Required]
-    public string Email { get; set; }
-    [Required]
-    public string Password { get; set; }
-}
