--- app/backend/Cower/Presentation/Models/LoginResponseDTO.cs
@@ -1,5 +1,0 @@
-namespace Cower.Presentation.Models;
-
-public record LoginResponseDTO(
-    UserResponseDTO User,
-    string Jwt);
