--- app/backend/Cower/Presentation/Models/RegisterResponseDTO.cs
@@ -1,5 +1,0 @@
-namespace Cower.Presentation.Models;
-
-public record RegisterResponseDTO(
-    UserResponseDTO User,
-    string Jwt);
