--- app/backend/Cower/Presentation/Models/UserInfoResponseDTO.cs
@@ -1,3 +1,0 @@
-namespace Cower.Presentation.Models;
-
-public record UserInfoResponseDTO(UserResponseDTO User);
