--- app/backend/Cower/Presentation/Models/UserResponseDTO.cs
@@ -1,9 +1,0 @@
-namespace Cower.Presentation.Models;
-
-public record UserResponseDTO(
-    long Id,
-    string Email,
-    string Role,
-    string? Name,
-    string? Surname,
-    string? Phone);
