--- app/backend/Cower/Service/IJwtService.cs
@@ -1,8 +1,0 @@
-using Cower.Data.Models;
-
-namespace Cower.Service;
-
-public interface IJwtService
-{
-    public string GenerateJwt(User user);
-}
