--- app/backend/Cower/Service/Models/RegisterUserRequestBL.cs
@@ -1,10 +1,0 @@
-using Cower.Data.Models;
-
-namespace Cower.Service.Models;
-
-public record RegisterUserRequestBL(
-    string Email,
-    string Password,
-    string? Name,
-    string? Surname,
-    string? Phone);
